Update logic for grading calculate button.

+++ GradingScale/GradingScale/GradingHelperViewController.swift
@@ -23,8 +23,8 @@ class GradingHelperViewController: UIViewController {
     @IBOutlet weak var TotalQuestionsTextField: UITextField!
     @IBOutlet weak var TotalQuestionsCorrectTextField: UITextField!
     @IBAction func CalculateButton(_ sender: UIButton) {
-        let TotalQuestions  = Int(TotalQuestionsTextField.text!)
-        let TotalQuestionsCorrect = Int(TotalQuestionsCorrectTextField.text!)
+        let TotalQuestions  = Double(TotalQuestionsTextField.text!)
+        let TotalQuestionsCorrect = Double(TotalQuestionsCorrectTextField.text!)
         let score = TotalQuestionsCorrect!/TotalQuestions!*100
         let scoreString = String(score)
         ScoreTextField.text = scoreString
